requirements.txt
fastapi
uvicorn
faiss-cpu
numpy
python-multipart
